Moving enhancer file to enhancer directory
