+++ src/workers/canCardMoveWorker.ts
@@ -1,6 +1,6 @@
 /* istanbul ignore file */
 import { expose } from "comlink";
-import { enhanceCard, enhanceSolitaire } from "lib/enhancers";
+import { enhanceCard, enhanceSolitaire } from "lib/enhancers/enhancers";
 import { columnFromLocation, finalFromLocation, makeCardLocationAware } from "lib/util";
 import { LocationAwareSolitaireCard, Solitaire } from "types/game";
 import { CanCardMoveFromWorker } from "types/worker";
