+++ src/workers/isGameCompleteWorker.ts
@@ -1,6 +1,6 @@
 /* istanbul ignore file */
 import { expose } from 'comlink';
-import { enhanceSolitaire } from 'lib/enhancers';
+import { enhanceSolitaire } from 'lib/enhancers/enhancers';
 import { Solitaire } from "types/game";
 
 const isGameComplete = (solitaire: Solitaire): boolean => {
